Use unprefixed constants in {set|get}_blocking. This fixes TermUX builds.
Many of lwIP's constant defines are guarded behind IFNDEFs, so depending on whether <fcntl.h> is included or not the values of those defines are different.
Apparently on regular Linux it's not included, but on TermUX it is included.

This also makes the constants exposed in Python bindings pretty much useless.